requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
streamlit>=1.28.0
plotly>=5.18.0

